stress-qsort: cast return from stress_mwc32() to match assigned type
Cast the 32 bit random value from stress_mwc32() to match the int32_t
type of *ptr to clean up implicit cast warnings.

+++ stress-qsort.c
@@ -142,7 +142,7 @@ static int stress_qsort(const stress_args_t *args)
 
 	/* This is expensive, do it once */
 	for (ptr = data, i = 0; i < n; i++)
-		*ptr++ = stress_mwc32();
+		*ptr++ = (int32_t)stress_mwc32();
 
 	stress_set_proc_state(args->name, STRESS_STATE_RUN);
